DecodeInterface can now return int8/16/32 and uint8/16/32.

+++ ext_test.go
@@ -109,7 +109,7 @@ func TestDecodeExtWithMap(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	wanted := map[string]interface{}{"I": uint64(42)}
+	wanted := map[string]interface{}{"I": int8(42)}
 	if !reflect.DeepEqual(got, wanted) {
 		t.Fatalf("got %#v, but wanted %#v", got, wanted)
 	}
